Resolving review comments
